requirements.txt
python-dotenv
llama-index
llama-index-embeddings-gemini
llama-index-llms-gemini
